moved seating chart modals in a new folder for organization. added modal div to outside of root div for better accessibility for modals
