+++ client/public/index.html
@@ -38,6 +38,7 @@
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
+    <div id="modal"></div>
     <div id="root"></div>
     <!--
       This HTML file is a template.
